📑 :: GetReserveLinkService
